+++ test/dummy/app/models/message.rb
@@ -9,4 +9,6 @@ class Message
   validates :content, presence: true, length: {maximum: 280}
   validates :subject, presence: true, exclusion: {in: %w[forbidden]}
   validates :status, presence: true
+
+  validate -> { errors.add(:subject, :custom) if subject == "invalid" }
 end